pyproject.toml
[tool.poetry]
name = "user-management"
version = "0.1.0"
description = ""
authors = ["KateM_zz <zupperu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
ruff = "^0.0.291"
pre-commit = "^3.4.0"
httpx = "^0.25.1"
black = "^23.11.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
pytest-asyncio = "^0.21.1"
uvicorn = "^0.24.0.post1"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^5.0.1"
aioredis = "^2.0.1"
python-multipart = "^0.0.6"

[tool.black]
target-version = ['py311']
include = '\.pyi?$'
line-length = 100

[tool.isort]
profile = "black"
filter_files = true
line_length = 100

[tool.ruff]
line-length = 100
exclude = [
    "**/migrations/*.py"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
